tests added

+++ models/OrderModel.go
@@ -17,10 +17,10 @@ type OrderRepo struct {
 	conn *gorm.DB
 }
 
-//go:generate mockgen -destination=../mocks/orderModel_mock.go -package=mocks go-mux/services IStores
+//go:generate mockgen -destination=../mocks/orderModel_mock.go -package=mocks DIY2/models IOrderRepo
 type IOrderRepo interface {
 	BuyProduct(orderModel *OrderModel) error
-	GetTopOrders(orderModel *OrderModel, interval int64, limit int) ([]int64, error)
+	GetTopOrders(orderModel *OrderModel, interval int, limit int) ([]int64, error)
 }
 
 func NewOrderRepo(conn *gorm.DB) *OrderRepo {
@@ -32,7 +32,7 @@ func (m *OrderRepo) BuyProduct(orderModel *OrderModel) error {
 	return result.Error
 }
 
-func (m *OrderRepo) GetTopOrders(orderModel *OrderModel, interval int64, limit int) ([]int64, error) {
+func (m *OrderRepo) GetTopOrders(orderModel *OrderModel, interval int, limit int) ([]int64, error) {
 	var products []int64
 	whereQuery := fmt.Sprintf("sold_at>=now()-interval '%v hour' and STORE_ID = %v", interval, orderModel.StoreId)
 	err := m.conn.Select("product_id").Group("product_id").Where(whereQuery).Order("count(product_id) desc").Limit(limit).Table("orders").Find(&products).Error
